requirements.txt
dbt-duckdb==1.7.0
airflow-provider-duckdb
pandas